% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_idw_plot_labels.R
\name{add_map_labels}
\alias{add_map_labels}
\title{Add labels to IDW plots}
\usage{
add_map_labels(
  x,
  lab.select = c("islands", "bathymetry", "convention line", "mainland", "peninsula"),
  region = NA,
  new.places = NULL,
  lab.replace = FALSE,
  transform.new.places = TRUE,
  add.scale.bar = "br"
)
}
\arguments{
\item{x}{A list containing the ggplot object returned by \code{make_idw_map()} or \code{change_fill_color()}. Alternatively, a gg object.}

\item{lab.select}{A vector indicating which types of labels to include on the map.}

\item{region}{Vector indicating the plotting region. Only specify if x is a gg object.}

\item{new.places}{A data frame containing new locations and labels to add to the map.}

\item{lab.replace}{Logical indicating whether new.places should be added to the default labels or replace the default labels.}

\item{transform.new.places}{Logical indicating whether the data frame passed to new.places should be transformed to match the map projection.}

\item{add.scale.bar}{A character vector indicating where to add the scale bar. "br" for bottom right, "tl" for top left, etc. Default = "br"}
}
\value{
Returns input list with IDW plot labels added.
}
\description{
Select which place labels should be included on the map based on category and region. Currently supports regions: 'bs.all' and 'bs.south'
See data/placenames for the default list of locations.
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
