% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform_normal.R
\name{backtransform_normal}
\alias{backtransform_normal}
\title{Backtransformation of normal score transform}
\usage{
backtransform_normal(scores, z_score, tails = "none", make_plot = TRUE)
}
\arguments{
\item{scores}{Numeric vector of scores to be transformed}

\item{z_score}{List returned by akgfmaps::normal_transform().}

\item{tails}{Character vector indicating which transformation/extrapoltion method to use. "none" (default): No extrapolation. Scores outside the original range get transformed to min or max.; "equal": Calculate magnitude of standard deviations around the original mean, and extrapolate linearly.; "separate": Calculates a new standard deviation for values above and below the mean, then extrpolate.}
}
\description{
Backtransformation of normal score transform.
}
