% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cold_pool_area.R
\name{calculate_cold_pool_area}
\alias{calculate_cold_pool_area}
\title{Calculate cold pool area in the southeastern Bering Sea using varying interpolation methods}
\usage{
calculate_cold_pool_area(
  dat,
  dat.year,
  var.col,
  lat.col,
  lon.col,
  in.crs = "+proj=longlat +datum=NAD83",
  interpolation.crs,
  cell.resolution,
  nm = Inf,
  pre = NA
)
}
\arguments{
\item{dat}{Input data frame. Must include columns for latitude, longitde, and variable to be interpolated.}

\item{dat.year}{Year}

\item{var.col}{Character vector denoting the name of the variable column.}

\item{lat.col}{Character vector indicating the name of the latitude column.}

\item{lon.col}{Character vector indicating the name of the longitude column.}

\item{in.crs}{Character vector (PROJ4 or WKT2/PROJ6 string) indicating the coordinate reference system for the data.}

\item{interpolation.crs}{Character vector (PROJ4 or WKT2/PROJ6 string) indicating the coordinate reference system to use for the interpolation. Historically EPSG:3338}

\item{cell.resolution}{Dimensions of interpolation grid cell in meters.}

\item{nm}{Maximum number of cells to use for interpolation.}

\item{pre}{Prefix for file names in output (in development.)}
}
\description{
Compare
}
