% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_text_size.R
\name{change_text_size}
\alias{change_text_size}
\title{Function to change the size of text in plots}
\usage{
change_text_size(x, size.mult, scale.theme = 1.125)
}
\arguments{
\item{x}{ggplot object that requires text resizing}

\item{size.mult}{Size multiplier for rescaling text.}

\item{scale.theme}{Size multiplier for axis text, legend text, legend title}
}
\description{
Function to change the size of text in plots
}
\author{
Sean K. Rohan, \email{sean.rohan@noaa.gov}
}
