% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa_from_raster.R
\name{cpa_from_raster}
\alias{cpa_from_raster}
\title{Calculate cold pool area from a raster}
\usage{
cpa_from_raster(x, raster_units = "m", temperature_threshold)
}
\arguments{
\item{x}{A RasterLayer object.}

\item{raster_units}{Character vector indicating units for x and y dimensions in the raster. Default = "m" for meters.}

\item{temperature_threshold}{Numeric. Temperature threshold for value.}
}
\description{
Calculate area of raster cells less than or equal to a temperature threshold.
}
