% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_file.R
\name{create_map_file}
\alias{create_map_file}
\title{Create map file}
\usage{
create_map_file(
  in.dat,
  file.prefix = NA,
  file.path = NA,
  try.change_text_size = FALSE,
  width = 12,
  height = 9,
  units = "in",
  res = 300,
  bg = "transparent"
)
}
\arguments{
\item{in.dat}{List containing ggplot object}

\item{file.prefix}{Character vector indicating what name should precede the file name. If not provided, uses \code{key.title} from the output of \code{make_idw_map}}

\item{file.path}{File path where the png should be saved. By default, Saves to working directory.}

\item{try.change_text_size}{Logical indicating whether labels of the input list should be resized using predefined scaling factors.}

\item{width}{Width of the .png. Passed to \code{png}.}

\item{height}{Height of the .png. Passed to \code{png}.}

\item{units}{Units of width and height. Passed to \code{png}.}

\item{res}{Resolution of the output. Passed to \code{png}.}

\item{bg}{Color of the background. Passed to \code{png}.}
}
\description{
Creates a .png map with the assigned file prefix in the designated file path (i.e. directory)
}
