% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_layer_guide.R
\name{generate_layer_guide}
\alias{generate_layer_guide}
\title{Generate a pdf file showing layers}
\usage{
generate_layer_guide(out_loc = "akgfmaps_layer_guide.pdf")
}
\arguments{
\item{out_loc}{Filepath for output. Must be a pdf.}
}
\description{
Generates a pdf file with plots of layers, file names, and default coordinate reference systems for all of the shapefiles that are included in the package.
}
