% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_region_guide.R
\name{generate_region_guide}
\alias{generate_region_guide}
\title{Generate blank region catalogue}
\usage{
generate_region_guide(
  out_loc = "akgfmaps_region_layers.pdf",
  select.regions = NULL
)
}
\arguments{
\item{out_loc}{Filepath for output. Must be a pdf.}

\item{select.regions}{Which regions to include in the output? Default selects all supported regions.}
}
\description{
Generates a pdf file with plots of layers returned by get_base_layers().
}
