% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_region_zip.R
\name{generate_region_zip}
\alias{generate_region_zip}
\title{Save regional shapefiles to a zip file}
\usage{
generate_region_zip(select.region, set.crs = "EPSG:4326", zip.path = NULL)
}
\arguments{
\item{select.region}{Character vector indicating which region. Options = ebs or bs.all, sebs or bs.south, nbs or bs.north, ecs, ebs.ecs, ai, ai.west, ai.central, ai.east, goa, goa.west, goa.east, ebs.slope, bssa1, bssa2, bssa3, bssa4, bssa5, bssa6}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', an Albers Equal Area coordinate reference system is automatically assigned.}

\item{zip.path}{Path to the zip file where shapefiles should be written.}
}
\description{
Write regional shapefiles returned by get_base_layers() and a README file with version information to a zip file.
}
\examples{
generate_region_zip(select.region = "ebs", set.crs = "WGS84")
}
