% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adfg_areas.R
\name{get_adfg_areas}
\alias{get_adfg_areas}
\title{Alaska Department of Fish and Game (ADFG) management areas}
\usage{
get_adfg_areas(set.crs = "auto", subset.fields = TRUE)
}
\arguments{
\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area coordinate reference system (EPSG:3338) is automatically assigned.}
}
\value{
ADFG management area sf POLYGON.
}
\description{
Retrieve ADFG management area POLYGON grid.
}
