% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_layers.R
\name{get_base_layers}
\alias{get_base_layers}
\title{Function to get base layers for bottom trawl and longline survey regions, bathymetry, and land}
\usage{
get_base_layers(
  select.region,
  design.year = NULL,
  set.crs = "EPSG:4269",
  split.land.at.180 = TRUE,
  fix.invalid.geom = TRUE,
  high.resolution.coast = FALSE
)
}
\arguments{
\item{select.region}{Character vector indicating which region to retrieve. Bottom trawl survey options: ebs or bs.all, sebs or bs.south, nbs or bs.north, ecs, ai, ai.west, ai.central, ai.east, goa, goa.west, goa.east, ebs.slope, bssa1, bssa2, bssa3, bssa4, bssa5, bssa6. Longline survey options: ll.ebs, ll.bssa1, ll.bssa2, ll.bssa3, ll.bssa4, ll.bssa5, ll.ai, ll.ai.west, ll.ai.central, ll.goa, ll.goa.west, ll.goa.central, ll.goa.east}

\item{design.year}{Survey design year for files to retrieve. Returns the most recent year when NULL or closest year before the design year when there isn't an exact match.}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area (EPSG:3338) is used.}

\item{split.land.at.180}{Logical. If set.crs is a geographic coordinate system, should the land polygon be split at 180 degrees to prevent polygons from wrapping around the world? Default = TRUE.}

\item{fix.invalid.geom}{Should invalid geometries be corrected using st_make_valid() and st_wrap_dateline()?}

\item{high.resolution.coast}{Should the State of Alaska polygon be a high resolution Alaska Department of Natural Resources 1:63360 scale polygon that includes smaller islands and a more detailed coastline? The higher resolution polygon (high.resolution.coast = TRUE) takes longer to load/plot and is recommended for spatial operations performed at high resolution (e.g., masking high resolution rasters). The lower resolution polygon (high.resolution.coast = FALSE) is recommended for general mapping and visualization purposes. Default = FALSE.}
}
\value{
A list containing sf objects land, bathymetry, survey area boundary, survey strata, survey grid (or survey stations for the longline survey), a data frame of feature labels (for the EBS shelf bottom trawl survey), coordinate reference system, plot boundaryies, axis label breaks, and graticule.
}
\description{
This function retrieves layers that are commonly used for mapping and spatial analysis of AFSC bottom trawl survey and longline survey data.
}
\examples{
\dontrun{
library(akgfmaps)

# EBS bottom trawl survey layers in Alaska Albers Equal Area projection (EPSG:3338)
sebs <- get_base_layers(select.region = "sebs",
                        set.crs = "EPSG:3338")

ggplot() +
  geom_sf(data = sebs$akland) +
  geom_sf(data = sebs$survey.strata,
          fill = NA,
          mapping = aes(color = "Survey strata")) +
  geom_sf(data = sebs$survey.grid,
          fill = NA,
          mapping = aes(color = "Station grid")) +
  geom_sf(data = sebs$survey.area,
          fill = NA,
          mapping = aes(color = "Survey area")) +
  geom_sf(data = sebs$graticule, alpha = 0.3, linewidth = 0.5) +
  scale_x_continuous(limits = sebs$plot.boundary$x,
                     breaks = sebs$lon.breaks) +
  scale_y_continuous(limits = sebs$plot.boundary$y,
                     breaks = sebs$lat.breaks) +
  theme_bw()

# EBS BT survey layers in NAD83 (EPSG:4269) with corner stations, 2022 design year strata, 
# and high resolution coastline. High resolution coastline takes longer to load and plot but is 
# recommended when land polygons are used for spatial analysis.

sebs_corners <- get_base_layers(select.region = "sebs",
                                set.crs = "EPSG:4269",
                                design.year = 2022,
                                high.resolution.coast = TRUE)

ggplot() +
  geom_sf(data = sebs_corners$akland) +
  geom_sf(data = sebs_corners$survey.strata,
          fill = NA,
          mapping = aes(color = "Survey strata")) +
  geom_sf(data = sebs_corners$survey.grid,
          fill = NA,
          mapping = aes(color = "Station grid")) +
  geom_sf(data = sebs_corners$survey.area,
          fill = NA,
          mapping = aes(color = "Survey area")) +
  geom_sf(data = sebs_corners$graticule, alpha = 0.3, linewidth = 0.5) +
  scale_x_continuous(limits = sebs_corners$plot.boundary$x,
                     breaks = sebs_corners$lon.breaks) +
  scale_y_continuous(limits = sebs_corners$plot.boundary$y,
                     breaks = sebs_corners$lat.breaks) +
  theme_bw()

# EBS slope, Aleutian Islands, and Gulf of Alaska surveys have subarea options
ai_west <- get_base_layers(select.region = "ai.west",
                           set.crs = "EPSG:3338")

ggplot() +
  geom_sf(data = ai_west$akland) +
  geom_sf(data = ai_west$survey.strata,
          fill = NA,
          mapping = aes(color = "Survey strata")) +
  geom_sf(data = ai_west$survey.grid,
          fill = NA,
          mapping = aes(color = "Station grid")) +
  geom_sf(data = ai_west$survey.area,
          fill = NA,
          mapping = aes(color = "Survey area")) +
  geom_sf(data = ai_west$graticule, alpha = 0.3, linewidth = 0.5) +
  scale_x_continuous(limits = ai_west$plot.boundary$x,
                     breaks = ai_west$lon.breaks) +
  scale_y_continuous(limits = ai_west$plot.boundary$y,
                     breaks = ai_west$lat.breaks) +
  theme_bw()}
}
