% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crab_strata.R
\name{get_crab_strata}
\alias{get_crab_strata}
\title{Get crab stratum layers}
\usage{
get_crab_strata(select.stock = NULL, select.region = "ebs", set.crs)
}
\arguments{
\item{select.region}{Select region}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area coordinate reference system (EPSG:3338) is automatically assigned.}

\item{select.unit}{Character vector indicating the name of the stock(s) to return. Options are Bristol Bay RKC ('bbrkc'), Pribilof Islands RKC ('pirkc'), Pribilof Islands BKC ('pibkc'), St. Matthew's BKC ('smbkc'), Norton Sound RKC ('nsrkc'), eastern Bering Sea snow crab ('ebssc'), and eastern Bering Sea Tanner crab ('ebstc').}
}
\description{
Load crab stratum polygons for EBS/NBS red king crab (Bristol Bay, Pribilof Islands, Norton Sound), EBS blue king crab (Pribilof Islands, St. Matthew Island), EBS snow crab, and EBS Tanner crab (West of 166W, East of 166W). Species and areas are described below.
}
