% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inpfc_strata.R
\name{get_inpfc_strata}
\alias{get_inpfc_strata}
\title{INPFC strata for the GOA or AI}
\usage{
get_inpfc_strata(select.region, set.crs, design.year = NULL)
}
\arguments{
\item{select.region}{Region for INPFC strata. Either GOA or AI.}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', an Albers Equal Area coordinate reference system is automatically assigned.}

\item{design.year}{Survey design year for files to retrieve. Returns the most recent year when NULL or closest year before the design year when there isn't an exact match.}
}
\value{
INPFC strata as an sf POLYGON.
}
\description{
Create INPFC strata sf POLYGONs by dissolving GOA or AI bottom trawl survey strata POLYGONs based on numerical stratum codes.
}
\examples{
\dontrun{
library(akgfmaps)

# INPFC strata for the Aleutian Islands (within AI bottom trawl survey extent)
inpfc_ai <- get_inpfc_strata(select.region = "ai", set.crs = "EPSG:3338")

ggplot() +
  geom_sf(data = inpfc_ai,
          mapping = aes(fill = INPFC_STRATUM))

# INPFC strata for the Gulf of Alaska (within GOA bottom trawl survey extent)
inpfc_goa <- get_inpfc_strata(select.region = "goa", set.crs = "EPSG:3338")

ggplot() +
  geom_sf(data = inpfc_goa,
          mapping = aes(fill = INPFC_STRATUM))}
}
