% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_stations.R
\name{get_survey_stations}
\alias{get_survey_stations}
\title{Character vector of station ID codes}
\usage{
get_survey_stations(select.region)
}
\arguments{
\item{select.region}{Choice of region. Options are "ebs", "bs.all", "sebs", "bs.south"}
}
\value{
A chracter vector of RACE station ID codes.
}
\description{
Returns a character vector of station id codes in a region.
}
