% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_grid.R
\name{make_2d_grid}
\alias{make_2d_grid}
\title{Make a raster, point, or polygon grid from an sf polygon}
\usage{
make_2d_grid(
  obj,
  resolution = c(3704, 3704),
  output_type = "point",
  bbox = NULL,
  model = "semi-open"
)
}
\arguments{
\item{obj}{An sf polygon.}

\item{resolution}{Resolution for grid cells in meters. Default = c(3704, 3704), equivalent of 2x2 nautical miles.}

\item{output_type}{Type of output as a 1L character vector ("point", "raster", "polygon).}

\item{bbox}{Optional. A 4L numeric vector to define the edges of the interpolation grid c(xmin, ymin, xmax, ymax).}

\item{model}{Character vector indicating the geometry model to use for st_intersection boundaries. Default = "semi-open"; for details see ?s2::s2_options}
}
\description{
Creates a raster (terra 'SpatRaster'), point grid (sf 'POINT'), or polygon mesh/'fishnet' (sf 'POLYGON') from an sf polygon.
}
