% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wkt_table.R
\name{make_wkt_table}
\alias{make_wkt_table}
\title{Convert sf to a data.frame with WKT geometry}
\usage{
make_wkt_table(x, area.crs, wkt.crs = "WGS84")
}
\arguments{
\item{x}{A simple features objects}

\item{area.crs}{Character vector or numeric indicating which coordinate reference system to use for calculating areas. The default is Alaska Albers Equal Area ("EPSG:3338")}

\item{wkt.crs}{Character vector or numeric indicating which coordinate reference system to use for WKT outputs. The default is WGS84 ("WGS84")}
}
\description{
This function converts a simple features (sf) geometry to a data.frame containing well-known text(WKT) representation of vector geometries and calculates the area of POLYGONS.
}
