% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOODLES}
\alias{NOODLES}
\title{NBS noodle bryozoan CPUE data from 2017-2023}
\format{
A data frame with 832 rows and 7 columns:
\describe{
  \item{COMMON_NAME}{Species common name}
  \item{YEAR}{Survey year}
  \item{HAUL}{Haul number}
  \item{STATIONID}{Station ID}
  \item{LONGITUDE}{Longitude in decimal degrees}
  \item{LATITUDE}{Latitude in decimal degrees}
  \item{CPUE_KGHA}{Catch per unit effort in kilograms per hectare}
  ...
}
}
\source{
\url{https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center}
}
\usage{
NOODLES
}
\description{
Noodle bryozoan (Alcyonidium enteromorpha) CPUE data from the NOAA/AFSC 2017, 2019, and 2021-2023 summer bottom trawl survey of the northern Bering Sea.
}
\keyword{datasets}
