% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_transform.R
\name{normal_transform}
\alias{normal_transform}
\title{Normal score transform}
\usage{
normal_transform(x)
}
\arguments{
\item{x}{Numeric vector of values to transform}
}
\value{
A list containing a vector of Z scores and a data frame containing the sorted vector and associated Z-scores.
}
\description{
Normal score transformation of a vector to mean zero, variance one.
}
