% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_and_mask.R
\name{rasterize_and_mask}
\alias{rasterize_and_mask}
\title{Convert spatial data to raster and mask using a shapefile}
\usage{
rasterize_and_mask(sgrid, amask)
}
\arguments{
\item{sgrid}{Gridded spatial data}

\item{amask}{Object to use as an area mask (simple features object or one of many sp objects)}
}
\description{
Function converts gridded spatial data with coordinates (eg. SpatialGrid) to a raster, then applies a spatial mask.
}
