% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data_frame_crs.R
\name{transform_data_frame_crs}
\alias{transform_data_frame_crs}
\title{Transform data frame to new coordinate reference system}
\usage{
transform_data_frame_crs(
  x,
  coords = c("x", "y"),
  in.crs = "+proj=longlat",
  out.crs,
  only.new.coords = FALSE
)
}
\arguments{
\item{x}{Input data frame.}

\item{coords}{Vector of names of coordinate column with order longitude and latitude. Default c("x", "y")}

\item{in.crs}{Coordinate reference system for input data frame. Default = '+proj=longlat'}

\item{out.crs}{Coordinate reference system for data frame to return. Default = '+proj=aea +lat_1=55 +lat_2=65 +lat_0=50 +lon_0=-154 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs'}

\item{only.new.coords}{Should only the new coordinates be returned without the input data frame? Default = FALSE}
}
\value{
Returns the input data frame with coordinates converted, or a data frame containing just the coordinates, if \code{only.new.coords == TRUE}.
}
\description{
Transform data frame to new coordinate reference system
}
\author{
Sean K. Rohan, \email{sean.rohan@noaa.gov}
}
