% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFS2018}
\alias{YFS2018}
\title{EBS yellowfin sole CPUE data from 2018}
\format{
A data frame with 376 rows and 7 columns:
\describe{
  \item{COMMON_NAME}{Species common name}
  \item{YEAR}{Survey year}
  \item{HAUL}{Haul number}
  \item{STATIONID}{Station ID}
  \item{LONGITUDE}{Longitude in decimal degrees}
  \item{LATITUDE}{Latitude in decimal degrees}
  \item{CPUE_KGHA}{Catch per unit effort in kilograms per hectare}
  ...
}
}
\source{
\url{https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center}
}
\usage{
YFS2018
}
\description{
Yellowfin sole CPUE data from the NOAA/AFSC 2018 summer bottom trawl survey of the eastern Bering Sea continental shelf.
}
\keyword{datasets}
