% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akgfmaps-package.R
\docType{package}
\name{akgfmaps-package}
\alias{akgfmaps}
\alias{akgfmaps-package}
\title{akgfmaps: Alaska Groundfish, Crab, and Ecosystem Survey Area Mapping}
\description{
Functions and vector geometry layers ('shapefiles') for making maps and conducting spatial analyses in Alaska marine regions.
}
\author{
\strong{Maintainer}: Sean Rohan \email{sean.rohan@noaa.gov}

Authors:
\itemize{
  \item Zack Oyafuso \email{zack.oyafuso@noaa.gov}
  \item Emily Markowitz \email{emily.markowitz@noaa.gov}
  \item Matt Callahan \email{matt.callahan@noaa.gov}
  \item Jason Conner \email{jason.conner@noaa.gov}
  \item Katy Echave \email{katy.echave@noaa.gov}
  \item Emily Ryznar \email{emily.ryznar@noaa.gov}
}

Other contributors:
\itemize{
  \item Liz Dawson [contributor]
  \item Ned Laman \email{ned.laman@noaa.gov} [contributor]
}

}
\keyword{internal}
