% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_fill_color.R
\name{change_fill_color}
\alias{change_fill_color}
\title{Change color scale on plots}
\usage{
change_fill_color(x, new.scheme, show.plot = TRUE, ...)
}
\arguments{
\item{x}{A list containing the ggplot object returned by \code{make_idw_map()}}
}
\value{
Returns a ggplot object with the fill scale replaced
}
\description{
Function to quickly change color palette for plots
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
