% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bsierp_regions.R
\name{get_bsierp_regions}
\alias{get_bsierp_regions}
\title{Bering Sea Integrated Ecosystem Research Program (BSIERP) regions}
\usage{
get_bsierp_regions(set.crs)
}
\arguments{
\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area coordinate reference system (EPSG:3338) is automatically assigned.}
}
\description{
Retrieve BSIERP region polygons.
}
