% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crab_strata.R
\name{get_crab_strata}
\alias{get_crab_strata}
\title{Get crab stratum layers}
\usage{
get_crab_strata(select.unit, set.crs)
}
\arguments{
\item{select.unit}{Select a species and management unit as a character vector. Options are Bristol Bay RKC ('bbrkc'), Pribilof Islands RKC ('pirkc'), Pribilof Islands BKC ('pibkc'), St. Matthew's BKC ('smbkc'), Norton Sound RKC ('nsrkc'), eastern Bering Sea snow crab ('ebssc'), and eastern Bering Sea Tanner crab ('ebstc').}

\item{set.crs}{Which coordinate reference system should be used?}
}
\description{
Load crab strata polygons for red king crab, blue king crab, snow crab, or Tanner crab. Species and areas are described below.
}
