% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esr_regions.R
\name{get_esr_regions}
\alias{get_esr_regions}
\title{Alaska Ecosystem Status Report (ESR) large marine ecosystem regions}
\usage{
get_esr_regions(select.region = "esr_subarea", set.crs)
}
\arguments{
\item{select.region}{Character vector indicating which region to retrieve. Options are: 'esr_subarea', 'esr_area', 'esr_suberea_inside', 'esr_area_inside'}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area coordinate reference system (EPSG:3338) is automatically assigned.}
}
\description{
Retrieve Alaska Ecosystem Status Report area polygons.
}
\examples{
\dontrun{
library(akgfmaps)

# ESR subareas
esr_subareas <- get_esr_regions(select.region = "esr_subarea", set.crs = 3338)

ggplot() +
  geom_sf(data = esr_subareas,
          mapping = aes(fill = AREA_NAME))

# ESR regions
esr_areas <- get_esr_regions(select.region = "esr_area", set.crs = 3338)

ggplot() +
  geom_sf(data = esr_areas,
          mapping = aes(fill = AREA_NAME))

# ESR subareas in the Gulf of Alaska that include inside waters
esr_inside_subareas <- get_esr_regions(select.region = "esr_subarea_inside", set.crs = 3338)

ggplot() +
  geom_sf(data = esr_inside_subareas,
          mapping = aes(fill = AREA_NAME))
}
}
