% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esr_regions.R
\name{get_esr_regions}
\alias{get_esr_regions}
\title{Alaska Ecosystem Status Report (ESR) large marine ecosystem regions}
\usage{
get_esr_regions(set.crs)
}
\arguments{
\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area coordinate reference system (EPSG:3338) is automatically assigned.}
}
\description{
Retrieve Alaska Ecosystem Status Report area polygons.
}
