% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inpfc_strata.R
\name{get_inpfc_strata}
\alias{get_inpfc_strata}
\title{INPFC strata for the GOA or AI}
\usage{
get_inpfc_strata(select.region, set.crs)
}
\arguments{
\item{select.region}{Region for INPFC strata. Either GOA or AI.}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', an Albers Equal Area coordinate reference system is automatically assigned.}
}
\value{
INPFC strata as an sf POLYGON.
}
\description{
Create INPFC strata sf POLYGONs by dissolving GOA or AI bottom trawl survey strata POLYGONs based on numerical stratum codes.
}
