% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmfs_areas.R
\name{get_nmfs_areas}
\alias{get_nmfs_areas}
\title{NMFS Statistical Areas}
\usage{
get_nmfs_areas(set.crs)
}
\arguments{
\item{set.crs}{Which coordinate reference system should be used? If 'auto', an Albers Equal Area coordinate reference system is automatically assigned.}
}
\value{
NMFS statistical areas in Alaska as an sf object.
}
\description{
Load NMFS statistical areas shapefile.
}
