% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_bathymetry.R
\name{get_survey_bathymetry}
\alias{get_survey_bathymetry}
\title{Function to get survey bathymetry layers}
\usage{
get_survey_bathymetry(select.region, set.crs)
}
\arguments{
\item{select.region}{Character vector indicating which region to retrieve. Options = ebs or bs.all, sebs or bs.south, nbs or bs.north}

\item{set.crs}{Which coordinate reference system should be used? If 'auto', Alaska Albers Equal Area (EPSG:3338) will be used.}
}
\value{
Returns an sf multiline object with survey contours.
}
\description{
Retrieves bathymetric contours that approximate groundfish bottom trawl survey stratum boundaries on the EBS shelf.
}
