% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_idw_map.R
\name{make_idw_map}
\alias{make_idw_map}
\title{Make IDW maps of CPUE for the EBS/NBS}
\usage{
make_idw_map(
  x = NA,
  COMMON_NAME = NA,
  LATITUDE = NA,
  LONGITUDE = NA,
  CPUE_KGHA = NA,
  region = "bs.south",
  extrap.box = NA,
  set.breaks = "jenks",
  grid.cell = c(0.05, 0.05),
  in.crs = "+proj=longlat",
  out.crs = "auto",
  key.title = "auto",
  log.transform = FALSE,
  idw.nmax = 4,
  use.survey.bathymetry = TRUE,
  return.continuous.grid = TRUE
)
}
\arguments{
\item{x}{Data frame which contains at minimum: CPUE, LATITUDE, and LONGITUDE. Can be passed as vectors instead (see below). Default value: \code{NA}}

\item{COMMON_NAME}{Common name}

\item{LATITUDE}{Latitude (degrees north)}

\item{LONGITUDE}{Longitude (degrees east; Western hemisphere is negative)}

\item{CPUE_KGHA}{Catch per unit effort in kilograms per hectare}

\item{region}{Character vector indicating which plotting region to use. Options: bs.south, bs.north, bs.all}

\item{extrap.box}{Optional. Vector specifying the dimensions of the extrapolation grid. Elements of the vector should be named to specify the minimum and maximum x and y values c(xmn, xmx, ymn, ymx). If not provided, region will be used to set the extrapolation area.}

\item{set.breaks}{Suggested. Vector of break points to use for plotting. Alternatively, a character vector indicating which break method to use. Default = "jenks"}

\item{grid.cell}{Optional. Numeric vector of length two, specifying the resolution for the extrapolation grid in degrees. Default c(0.05,0.05)}

\item{in.crs}{Character vector containing the coordinate reference system for projecting the extrapolation grid.}

\item{out.crs}{Character vector containing the coordinate reference system for projecting the extrapolation grid.}

\item{key.title}{Character vector which will appear in the legend above CPUE (kg/ha). Default = "auto" tries to pull COMMON_NAME from input.}

\item{log.transform}{Character vector indicating whether CPUE values should be log-transformed for IDW. Default = FALSE.}

\item{idw.nmax}{Maximum number of adjacent stations to use for interpolation. Default = 8}

\item{use.survey.bathymetry}{Logical indicating if historical survey bathymetry should be used instead of continuous regional bathymetry. Default = TRUE}

\item{return.continuous.grid}{If TRUE, also returns an extrapolation grid on a continuous scale.}

\item{idp}{Inverse distance weighting power. Default = 2}
}
\value{
Returns a list containing: 
(1) plot: a ggplot IDW map;
(2) extrapolation.grid: the extrapolation grid with estimated values on a discrete scale;
(3) continuous.grid: extrapolation grid with estimates on a continuous scale;
(4) region: the region;
(5) n.breaks: the number of level breaks;
(6) key.title: title for the legend;
(7) crs: coordinate reference system as a PROJ6 (WKT2:2019) string;
}
\description{
This function can be used to make inverse-distance-weighted plots for the eastern Bering Sea and northern Bering Sea
}
\author{
Sean Rohan \email{sean.rohan@noaa.gov}
}
