% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFS2017}
\alias{YFS2017}
\title{Yellowfin sole CPUE data}
\format{
A data frame with 376 rows and 5 variables:
\describe{
  \item{LONGITUDE}{Longitude in decimal degrees}
  \item{LATITUDE}{Latitude in decimal degrees}
  ...
}
}
\source{
\url{https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center}
}
\usage{
YFS2017
}
\description{
Yellowfin sole CPUE data from NOAA/AFSC 2017 summer bottom trawl survey of the eastern Bering Sea continental shelf.
}
\keyword{datasets}
